@@ -10,11 +10,11 @@
     "axios": "^1.3.3",
     "node-sass": "^7.0.1",
     "nouislider-react": "^3.4.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-accessible-accordion": "^3.3.5",
     "react-bootstrap": "^2.5.0",
     "react-countup": "^6.2.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-modal-video": "^1.2.8",
     "react-parallax": "^3.3.0",
     "react-router-dom": "^6.8.1",
